# recipeApp/server/docs/swagger.yaml
basePath: /
definitions:
  handler.favPostRequest:
    properties:
      recipeid:
        type: integer
      userid:
        type: integer
    type: object
  handler.mealPostRequest:
    properties:
      date:
        type: string
      mealtype:
        type: string
      recipeid:
        type: integer
      userid:
        type: integer
    type: object
...
host: localhost:5000
info:
  contact:
    name: Wes Ahearn, Seth Paul
  description: This is the server for the Mallow Recipe Application
  license:
    name: Apache 2.0
  title: Mallow Recipe Server API
  version: "1.0"
paths:
  /server/favorites/add:
    post:
      consumes:
      - application/json
      parameters:
      - description: favorite data
        in: body
        name: requestBody
        required: true
        schema:
          $ref: '#/definitions/handler.favPostRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
      summary: add a favorite item to the database
...
  /server/meals/add:
    post:
      consumes:
      - application/json
      parameters:
      - description: meal data
        in: body
        name: requestBody
        required: true
        schema:
          $ref: '#/definitions/handler.mealPostRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
      summary: post a mealplan item to the database
...
swagger: "2.0"